+++ crates/try_v2_derive/examples/pass_MultipleGenerics.rs
@@ -2,11 +2,12 @@
 #![feature(try_trait_v2)]
 #![feature(try_trait_v2_residual)]
 
-use try_v2_derive::{Try, Try_ConvertResult};
+use try_v2_derive::Try;
 
-#[derive(Debug, Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
-enum Exit<T, E> {
+pub enum Exit<T, E> {
     Ok(T),
     TestsFailed,
     OtherError(E),