+++ crates/try_v2_derive/tests/compilation/fail_FirstGenericNotOutput.rs
@@ -2,23 +2,25 @@
 #![feature(try_trait_v2)]
 #![feature(try_trait_v2_residual)]
 
-use try_v2_derive::{Try, Try_ConvertResult};
+use try_v2_derive::Try;
 
-#[derive(Try, Try_ConvertResult)]
-#[must_use]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 enum Owned<T, E> {
     Ok(E),
     Err(T),
 }
 
-#[derive(Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
 enum Borrowed<'t, 'e, T, E> {
     Ok(&'e E),
     Err(&'t T),
 }
 
-#[derive(Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
 enum MultipleBorrowed<'t, 'e, 'f, T, E, F> {
     Ok(&'e E, &'f F),