+++ crates/try_v2_derive/tests/compilation/fail_OutputNamedFields.rs
@@ -2,17 +2,18 @@
 #![feature(try_trait_v2)]
 #![feature(try_trait_v2_residual)]
 
-use try_v2_derive::{Try, Try_ConvertResult};
+use try_v2_derive::Try;
 
-#[derive(Try, Try_ConvertResult)]
-#[must_use]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 enum OutputNamedField<T> {
     Ok { foo: T },
     TestsFailed(T),
     OtherError(String),
 }
 
-#[derive(Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
 enum OutputNamedFieldBorrowed<'t, T> {
     Ok { foo: &'t T },