+++ crates/try_v2_derive/tests/compilation/fail_TooManyOutputs.rs
@@ -2,23 +2,25 @@
 #![feature(try_trait_v2)]
 #![feature(try_trait_v2_residual)]
 
-use try_v2_derive::{Try, Try_ConvertResult};
+use try_v2_derive::Try;
 
-#[derive(Try, Try_ConvertResult)]
-#[must_use]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 enum TooManyOutputs<T, E> {
     Ok(T, E),
     Err,
 }
 
-#[derive(Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
 enum TooManyOutputsBorrowed<'t, 'e, T, E> {
     Ok(&'t T, &'e E),
     Err,
 }
 
-#[derive(Try, Try_ConvertResult)]
+#[derive(Debug, Try)]
+#[FromResidual(Result<_, Self::Residual>)]
 #[must_use]
 enum TooManyOutputsBorrowedOrdering<'e, 't, T, E> {
     Ok(&'e E, &'t T),